+++ campus-second-goods-traning/src/main/java/com/campus/service/Impl/TradeServicelmpl.java
@@ -2,11 +2,12 @@ package com.campus.service.Impl;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.campus.entity.Trade;
 import com.campus.entity.TradeCreateDTO;
 import com.campus.entity.TradePageQueryDTO;
 import com.campus.mapper.TradeMapper;
-import com.campus.util.CurrentHolder;
+import com.campus.utils.CurrentHolder;
 import com.campus.utils.PageResult;
 import com.campus.service.TradeService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +16,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 
 @Service
-public class TradeServicelmpl implements TradeService {
+public class TradeServicelmpl  implements TradeService {
     @Autowired
     private TradeMapper tradeMapper;
 
@@ -33,13 +34,13 @@ public class TradeServicelmpl implements TradeService {
 
         trade.setProductId(tradeCreateDTO.getId());
 
-        Integer currentUserId = CurrentHolder.getCurrentId();
+        Long currentUserId = CurrentHolder.getCurrentId();
         // 添加空值检查
         if (currentUserId == null) {
             throw new IllegalStateException("无法获取当前用户ID");
         }
 
-        trade.setBuyerId(currentUserId.longValue());
+        trade.setBuyerId(currentUserId);
         trade.setSellerId(tradeCreateDTO.getSeller_id());
         trade.setProductImage(tradeCreateDTO.getImage_url());
         trade.setProductPrice(tradeCreateDTO.getPrice());
@@ -57,12 +58,12 @@ public class TradeServicelmpl implements TradeService {
     }
 
     @Override
-    public PageResult pageQuery(TradePageQueryDTO tradePageQueryDTO,Integer currentUserId) {
+    public PageResult pageQuery(TradePageQueryDTO tradePageQueryDTO,Long currentUserId) {
         Page<Trade> page = new Page<>(tradePageQueryDTO.getPage(), tradePageQueryDTO.getPageSize());
 
         System.out.println("执行分页查询，参数: " + tradePageQueryDTO);
 
-        IPage<Trade> tradePage = tradeMapper.pageQuery(page, tradePageQueryDTO,currentUserId.longValue());
+        IPage<Trade> tradePage = tradeMapper.pageQuery(page, tradePageQueryDTO,currentUserId);
 
         System.out.println("查询结果总数: " + tradePage.getTotal());
         System.out.println("查询结果记录数: " + tradePage.getRecords().size());
